chore: clean up environment variables
