+++ config.template.js
@@ -1,5 +1,5 @@
 window.env = {
-  ENV: '$ENV',
+  NAMESPACE: '$NAMESPACE',
   OIDC_ISSUER: '$OIDC_ISSUER',
   FDK_REGISTRATION_BASE_URI: '$FDK_REGISTRATION_BASE_URI',
   TERMS_AND_CONDITIONS_HOST: '$TERMS_AND_CONDITIONS_HOST',
