+++ docker-compose.yml
@@ -6,9 +6,10 @@ services:
     ports:
       - 8181:8080
     environment:
-      - ENV
-      - OIDC_ISSUER
-      - FDK_REGISTRATION_BASE_URI
-      - TERMS_AND_CONDITIONS_HOST
-      - SEARCH_HOST
-      - ADMIN_GUI_HOST
+      - NAMESPACE=development
+      - OIDC_ISSUER=https://sso.staging.fellesdatakatalog.digdir.no/auth/realms/fdk
+      - FDK_REGISTRATION_BASE_URI=https://registrering.staging.fellesdatakatalog.digdir.no
+      - TERMS_AND_CONDITIONS_HOST=https://terms.staging.fellesdatakatalog.digdir.no
+      - SEARCH_HOST=https://staging.fellesdatakatalog.digdir.no
+      - ADMIN_GUI_HOST=https://admin.staging.fellesdatakatalog.digdir.no
+      - SKE_THEME_PROFILE=910244132
