+++ src/utils/common/index.ts
@@ -1,0 +1,17 @@
+import type { EnvironmentVariables } from '../../types';
+
+function assertIsDefined<T>(
+  key: string,
+  value: T
+): asserts value is NonNullable<T> {
+  if (value === undefined || value === null) {
+    throw new Error(`Expected ${key} to be defined, but received ${value}`);
+  }
+}
+
+export const validateEnv = (
+  env: EnvironmentVariables
+): EnvironmentVariables => {
+  Object.entries(env).forEach(([key, value]) => assertIsDefined(key, value));
+  return env;
+};
